refactor: move implementation of assertions for `CString`/`CStr` and `OsString`/OsStr` to own separate modules
